Moved createPubSubTopics to pubsub.ts
