+++ src/index-utils.ts
@@ -590,14 +590,3 @@ async function deleteFunction(projectId: string, functionName: string): Promise<
     console.log(`Function '${functionName}' not found or location not available.`);
   }
 }
-
-export async function createPubSubTopics(pubSubTopics: string[]) {
-  for (const topicName of pubSubTopics) {
-    const pubSubTopicRef = db.doc(`@topics/${topicName}`);
-    const pubSubTopic = await pubSubTopicRef.get();
-    if (pubSubTopic.exists) {
-      continue;
-    }
-    await pubSubTopicRef.set({timestamp: new Date()});
-  }
-}
