+++ src/tests/utils/pubsub.test.ts
@@ -4,7 +4,7 @@ import {initializeEmberFlow} from "../../index";
 import {dbStructure, Entity} from "../../sample-custom/db-structure";
 import {securityConfig} from "../../sample-custom/security";
 import {validatorConfig} from "../../sample-custom/validators";
-import {cleanPubSubProcessedIds, pubsubUtils} from "../../utils/pubsub";
+import {cleanPubSubProcessedIds, createPubSubTopics, pubsubUtils} from "../../utils/pubsub";
 import {ScheduledEvent} from "firebase-functions/lib/v2/providers/scheduler";
 import * as misc from "../../utils/misc";
 import {firestore} from "firebase-admin";
@@ -25,6 +25,41 @@ admin.initializeApp({
   databaseURL: "https://test-project.firebaseio.com",
 });
 initializeEmberFlow(projectConfig, admin, dbStructure, Entity, securityConfig, validatorConfig, []);
+
+describe("createPubSubTopics", () => {
+  let docSetMock: jest.Mock;
+  let docGetMock: jest.Mock;
+
+  beforeEach(() => {
+    docSetMock = jest.fn();
+    docGetMock = jest.fn();
+    const dbDoc = ({
+      set: docSetMock,
+      get: docGetMock,
+    } as unknown) as admin.firestore.DocumentReference<admin.firestore.DocumentData>;
+    jest.spyOn(admin.firestore(), "doc").mockReturnValue(dbDoc);
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
+  it("should create pubsub topics when not existing", async () => {
+    docGetMock.mockResolvedValue({exists: false});
+    await createPubSubTopics(["SAMPLE_TOPIC_NAME", "ANOTHER_TOPIC_NAME"]);
+
+    expect(docGetMock).toHaveBeenCalledTimes(2);
+    expect(docSetMock).toHaveBeenCalledTimes(2);
+  });
+
+  it("should skip creating pubsub topics when existing", async () => {
+    docGetMock.mockResolvedValue({exists: true});
+    await createPubSubTopics(["SAMPLE_TOPIC_NAME", "ANOTHER_TOPIC_NAME"]);
+
+    expect(docGetMock).toHaveBeenCalledTimes(2);
+    expect(docSetMock).toHaveBeenCalledTimes(0);
+  });
+});
 
 describe("pubsubUtils", () => {
   let colGetMock: jest.Mock;
